+++ .gitignore
@@ -1,3 +1,3 @@
 build
 __pycache__
-.vscode
+.cache
