+++ .vscode/.cortex-debug.peripherals.state.json
@@ -1,0 +1,1 @@
+[]
