+++ .vscode/settings.json
@@ -1,0 +1,22 @@
+{
+    // These settings tweaks to the cmake plugin will ensure
+    // that you debug using cortex-debug instead of trying to launch
+    // a Pico binary on the host
+    "cmake.statusbar.advanced": {
+        "debug": {
+            "visibility": "hidden"
+        },
+        "launch": {
+            "visibility": "hidden"
+        },
+        "build": {
+            "visibility": "hidden"
+        },
+        "buildTarget": {
+            "visibility": "hidden"
+        }
+    },
+    "cmake.buildBeforeRun": true,
+    "cmake.configureArgs": ["-DCMAKE_BUILD_TYPE=Debug"],
+    "C_Cpp.default.configurationProvider": "ms-vscode.cmake-tools"
+}
