+++ CMakeLists.txt
@@ -12,7 +12,7 @@ set(CMAKE_CXX_STANDARD 17)
 # Pull in Pico SDK (must be before project)
 include(pico_sdk_import.cmake)
 
-project(sensor C CXX ASM)
+project(main C CXX ASM)
 
 if (PICO_DEOPTIMIZED_DEBUG)
     message("Using fully de-optimized debug build (set PICO_DEOPTIMIZED_DEBUG=0 to optimize)")
@@ -29,20 +29,20 @@ pico_sdk_init()
 
 # Add executable. Default name is the project name, version 0.1
 
-add_executable(sensor)
-target_sources(sensor PRIVATE sensor.c i2c.c linalg.c sine_analysis.c adc.c pwm.c)
+add_executable(main)
+target_sources(main PRIVATE main.c i2c.c sine.c adc.c pwm.c)
 
-pico_set_program_name(sensor "sensor")
-pico_set_program_version(sensor "0.1")
+pico_set_program_name(main "main")
+pico_set_program_version(main "0.1")
 
-pico_enable_stdio_uart(sensor 0)
-pico_enable_stdio_usb(sensor 1)
+pico_enable_stdio_uart(main 0)
+pico_enable_stdio_usb(main 1)
 
 # Add the standard library to the build
-target_link_libraries(sensor pico_stdlib pico_multicore)
+target_link_libraries(main pico_stdlib pico_multicore)
 
 # Add any user requested libraries
-target_link_libraries(sensor
+target_link_libraries(main
         hardware_pio
         hardware_timer
         hardware_clocks
@@ -52,5 +52,5 @@ target_link_libraries(sensor
         hardware_i2c
         )
 
-pico_add_extra_outputs(sensor)
+pico_add_extra_outputs(main)
