--- linalg.h
@@ -1,2 +1,0 @@
-double dot(double vector1[], double vector2[], int dimension);
-double norm(double vector[], int dimension);
