+++ pwm.c
@@ -6,7 +6,6 @@
 #include "hardware/pwm.h"  
 #include "hardware/gpio.h"
 #include "hardware/dma.h"
-#include "dmapwm.h"
 
 #include "pwm.h"
