+++ sine.h
@@ -1,24 +1,27 @@
-// pin at which sine wave will be emitted
-#define INPUT_SIGNAL_PIN 0
-
-// amount to bit shift sample position such that the sample rate is reduced
-// by a factor. explicitly, decreases the sample rate by a factor
-// of 2^SAMPLE_RATE_FACTOR_SHIFT. e.g. to decrease the sample rate by a factor of 8,
-// set this to 3.
-#define SAMPLE_RATE_FACTOR_SHIFT 2
+// frequency (in kHz) to set clock at
+#define CLK_KHZ 250000
+#define KHZ_TO_HZ 1000.0
 
 // saturation voltage in volts
 #define VDD 3.3
 
-// amplitude of input sine wave in volts
-#define AMPLITUDE 0.1
+#define PI 3.1415926535
 
-// frequency (in kHz) to set clock at
-#define CLK_KHZ 200000
-#define KHZ_TO_HZ 1000.0
+typedef struct {
+    double amplitude;
+    double frequency;
+    double phase;
+    double offset;
+} sine;
 
-// for calculating sine
-#define PI 3.14159265
+typedef struct {
+    double magnitude;
+    double phase;
+} complex;
 
-void pwm_interrupt_handler();
-void generate_sine_wave(uint32_t sine_frequency);
+uint32_t * generate_sine_table(int length, double amplitude);
+int highest_frequency_to_table_length(float frequency);
+void average_period(uint8_t samples[], uint8_t input_period[], uint8_t output_period[], int num_samples_per_period, int num_periods);
+void samples_to_voltages(uint8_t samples[], double voltages[], int num_samples);
+sine characterize(double sine_period[], double frequency, int num_samples);
+complex impedence(sine vin, sine vout, double Rf);
