--- sine.h.backup2
@@ -1,54 +1,0 @@
-#define SINE_TABLE_LENGTH 50
-
-uint8_t SINE_TABLE[] = {
-    50,
-    50,
-    50,
-    51,
-    51,
-    51,
-    51,
-    51,
-    51,
-    51,
-    51,
-    51,
-    52,
-    52,
-    51,
-    51,
-    51,
-    51,
-    51,
-    51,
-    51,
-    51,
-    51,
-    50,
-    50,
-    50,
-    50,
-    50,
-    49,
-    49,
-    49,
-    49,
-    49,
-    49,
-    49,
-    49,
-    49,
-    48,
-    48,
-    49,
-    49,
-    49,
-    49,
-    49,
-    49,
-    49,
-    49,
-    49,
-    50,
-    50,
-};
