--- sine.h.backup3
@@ -1,54 +1,0 @@
-#define SINE_TABLE_LENGTH 50
-
-uint8_t SINE_TABLE[] = {
-    50,
-    52,
-    54,
-    56,
-    57,
-    59,
-    60,
-    62,
-    63,
-    64,
-    64,
-    65,
-    65,
-    65,
-    65,
-    64,
-    64,
-    63,
-    62,
-    60,
-    59,
-    57,
-    56,
-    54,
-    52,
-    50,
-    48,
-    46,
-    44,
-    43,
-    41,
-    40,
-    38,
-    37,
-    36,
-    36,
-    35,
-    35,
-    35,
-    35,
-    36,
-    36,
-    37,
-    38,
-    40,
-    41,
-    43,
-    44,
-    46,
-    48,
-};
