--- sinetable.h
@@ -1,104 +1,0 @@
-#define SINE_TABLE_LENGTH 100
-
-uint32_t SINE_TABLE[] = {
-    100,
-    102,
-    104,
-    106,
-    108,
-    109,
-    111,
-    113,
-    115,
-    116,
-    118,
-    119,
-    121,
-    122,
-    123,
-    125,
-    126,
-    127,
-    127,
-    128,
-    129,
-    129,
-    130,
-    130,
-    130,
-    130,
-    130,
-    130,
-    130,
-    129,
-    129,
-    128,
-    127,
-    127,
-    126,
-    125,
-    123,
-    122,
-    121,
-    119,
-    118,
-    116,
-    115,
-    113,
-    111,
-    109,
-    108,
-    106,
-    104,
-    102,
-    100,
-    98,
-    96,
-    94,
-    92,
-    91,
-    89,
-    87,
-    85,
-    84,
-    82,
-    81,
-    79,
-    78,
-    77,
-    75,
-    74,
-    73,
-    73,
-    72,
-    71,
-    71,
-    70,
-    70,
-    70,
-    70,
-    70,
-    70,
-    70,
-    71,
-    71,
-    72,
-    73,
-    73,
-    74,
-    75,
-    77,
-    78,
-    79,
-    81,
-    82,
-    84,
-    85,
-    87,
-    89,
-    91,
-    92,
-    94,
-    96,
-    98,
-};
