--- sinetables.h
@@ -1,5540 +1,0 @@
-#define LENGTH_100_SINE_TABLE_LENGTH 100
-uint32_t LENGTH_100_SINE_TABLE[] = {
-    100,
-    104,
-    108,
-    111,
-    115,
-    119,
-    122,
-    126,
-    129,
-    132,
-    136,
-    139,
-    141,
-    144,
-    147,
-    149,
-    151,
-    153,
-    155,
-    156,
-    158,
-    159,
-    160,
-    160,
-    160,
-    161,
-    160,
-    160,
-    160,
-    159,
-    158,
-    156,
-    155,
-    153,
-    151,
-    149,
-    147,
-    144,
-    141,
-    139,
-    136,
-    132,
-    129,
-    126,
-    122,
-    119,
-    115,
-    111,
-    108,
-    104,
-    100,
-    96,
-    92,
-    89,
-    85,
-    81,
-    78,
-    74,
-    71,
-    68,
-    64,
-    61,
-    59,
-    56,
-    53,
-    51,
-    49,
-    47,
-    45,
-    44,
-    42,
-    41,
-    40,
-    40,
-    40,
-    39,
-    40,
-    40,
-    40,
-    41,
-    42,
-    44,
-    45,
-    47,
-    49,
-    51,
-    53,
-    56,
-    59,
-    61,
-    64,
-    68,
-    71,
-    74,
-    78,
-    81,
-    85,
-    89,
-    92,
-    96,
-};
-
-#define LENGTH_200_SINE_TABLE_LENGTH 200
-uint32_t LENGTH_200_SINE_TABLE[] = {
-    200,
-    204,
-    208,
-    211,
-    215,
-    219,
-    223,
-    226,
-    230,
-    234,
-    237,
-    241,
-    245,
-    248,
-    252,
-    255,
-    258,
-    262,
-    265,
-    268,
-    271,
-    274,
-    277,
-    280,
-    283,
-    286,
-    288,
-    291,
-    293,
-    296,
-    298,
-    300,
-    302,
-    304,
-    306,
-    308,
-    310,
-    311,
-    313,
-    314,
-    315,
-    316,
-    317,
-    318,
-    319,
-    320,
-    320,
-    321,
-    321,
-    321,
-    321,
-    321,
-    321,
-    321,
-    320,
-    320,
-    319,
-    318,
-    317,
-    316,
-    315,
-    314,
-    313,
-    311,
-    310,
-    308,
-    306,
-    304,
-    302,
-    300,
-    298,
-    296,
-    293,
-    291,
-    288,
-    286,
-    283,
-    280,
-    277,
-    274,
-    271,
-    268,
-    265,
-    262,
-    258,
-    255,
-    252,
-    248,
-    245,
-    241,
-    237,
-    234,
-    230,
-    226,
-    223,
-    219,
-    215,
-    211,
-    208,
-    204,
-    200,
-    196,
-    192,
-    189,
-    185,
-    181,
-    177,
-    174,
-    170,
-    166,
-    163,
-    159,
-    155,
-    152,
-    148,
-    145,
-    142,
-    138,
-    135,
-    132,
-    129,
-    126,
-    123,
-    120,
-    117,
-    114,
-    112,
-    109,
-    107,
-    104,
-    102,
-    100,
-    98,
-    96,
-    94,
-    92,
-    90,
-    89,
-    87,
-    86,
-    85,
-    84,
-    83,
-    82,
-    81,
-    80,
-    80,
-    79,
-    79,
-    79,
-    79,
-    79,
-    79,
-    79,
-    80,
-    80,
-    81,
-    82,
-    83,
-    84,
-    85,
-    86,
-    87,
-    89,
-    90,
-    92,
-    94,
-    96,
-    98,
-    100,
-    102,
-    104,
-    107,
-    109,
-    112,
-    114,
-    117,
-    120,
-    123,
-    126,
-    129,
-    132,
-    135,
-    138,
-    142,
-    145,
-    148,
-    152,
-    155,
-    159,
-    163,
-    166,
-    170,
-    174,
-    177,
-    181,
-    185,
-    189,
-    192,
-    196,
-};
-
-#define LENGTH_300_SINE_TABLE_LENGTH 300
-uint32_t LENGTH_300_SINE_TABLE[] = {
-    300,
-    304,
-    308,
-    311,
-    315,
-    319,
-    323,
-    327,
-    330,
-    334,
-    338,
-    342,
-    345,
-    349,
-    353,
-    356,
-    360,
-    363,
-    367,
-    370,
-    374,
-    377,
-    381,
-    384,
-    388,
-    391,
-    394,
-    397,
-    401,
-    404,
-    407,
-    410,
-    413,
-    416,
-    419,
-    422,
-    424,
-    427,
-    430,
-    433,
-    435,
-    438,
-    440,
-    442,
-    445,
-    447,
-    449,
-    451,
-    454,
-    456,
-    457,
-    459,
-    461,
-    463,
-    465,
-    466,
-    468,
-    469,
-    470,
-    472,
-    473,
-    474,
-    475,
-    476,
-    477,
-    478,
-    479,
-    479,
-    480,
-    480,
-    481,
-    481,
-    481,
-    482,
-    482,
-    482,
-    482,
-    482,
-    481,
-    481,
-    481,
-    480,
-    480,
-    479,
-    479,
-    478,
-    477,
-    476,
-    475,
-    474,
-    473,
-    472,
-    470,
-    469,
-    468,
-    466,
-    465,
-    463,
-    461,
-    459,
-    457,
-    456,
-    454,
-    451,
-    449,
-    447,
-    445,
-    442,
-    440,
-    438,
-    435,
-    433,
-    430,
-    427,
-    424,
-    422,
-    419,
-    416,
-    413,
-    410,
-    407,
-    404,
-    401,
-    397,
-    394,
-    391,
-    388,
-    384,
-    381,
-    377,
-    374,
-    370,
-    367,
-    363,
-    360,
-    356,
-    353,
-    349,
-    345,
-    342,
-    338,
-    334,
-    330,
-    327,
-    323,
-    319,
-    315,
-    311,
-    308,
-    304,
-    300,
-    296,
-    292,
-    289,
-    285,
-    281,
-    277,
-    273,
-    270,
-    266,
-    262,
-    258,
-    255,
-    251,
-    247,
-    244,
-    240,
-    237,
-    233,
-    230,
-    226,
-    223,
-    219,
-    216,
-    212,
-    209,
-    206,
-    203,
-    199,
-    196,
-    193,
-    190,
-    187,
-    184,
-    181,
-    178,
-    176,
-    173,
-    170,
-    167,
-    165,
-    162,
-    160,
-    158,
-    155,
-    153,
-    151,
-    149,
-    146,
-    144,
-    143,
-    141,
-    139,
-    137,
-    135,
-    134,
-    132,
-    131,
-    130,
-    128,
-    127,
-    126,
-    125,
-    124,
-    123,
-    122,
-    121,
-    121,
-    120,
-    120,
-    119,
-    119,
-    119,
-    118,
-    118,
-    118,
-    118,
-    118,
-    119,
-    119,
-    119,
-    120,
-    120,
-    121,
-    121,
-    122,
-    123,
-    124,
-    125,
-    126,
-    127,
-    128,
-    130,
-    131,
-    132,
-    134,
-    135,
-    137,
-    139,
-    141,
-    143,
-    144,
-    146,
-    149,
-    151,
-    153,
-    155,
-    158,
-    160,
-    162,
-    165,
-    167,
-    170,
-    173,
-    176,
-    178,
-    181,
-    184,
-    187,
-    190,
-    193,
-    196,
-    199,
-    203,
-    206,
-    209,
-    212,
-    216,
-    219,
-    223,
-    226,
-    230,
-    233,
-    237,
-    240,
-    244,
-    247,
-    251,
-    255,
-    258,
-    262,
-    266,
-    270,
-    273,
-    277,
-    281,
-    285,
-    289,
-    292,
-    296,
-};
-
-#define LENGTH_400_SINE_TABLE_LENGTH 400
-uint32_t LENGTH_400_SINE_TABLE[] = {
-    400,
-    404,
-    408,
-    411,
-    415,
-    419,
-    423,
-    427,
-    430,
-    434,
-    438,
-    442,
-    445,
-    449,
-    453,
-    457,
-    460,
-    464,
-    468,
-    471,
-    475,
-    479,
-    482,
-    486,
-    489,
-    493,
-    496,
-    500,
-    503,
-    507,
-    510,
-    513,
-    517,
-    520,
-    523,
-    527,
-    530,
-    533,
-    536,
-    539,
-    542,
-    546,
-    549,
-    552,
-    555,
-    557,
-    560,
-    563,
-    566,
-    569,
-    571,
-    574,
-    577,
-    579,
-    582,
-    584,
-    587,
-    589,
-    592,
-    594,
-    596,
-    598,
-    601,
-    603,
-    605,
-    607,
-    609,
-    611,
-    612,
-    614,
-    616,
-    618,
-    619,
-    621,
-    622,
-    624,
-    625,
-    627,
-    628,
-    629,
-    631,
-    632,
-    633,
-    634,
-    635,
-    636,
-    637,
-    637,
-    638,
-    639,
-    639,
-    640,
-    641,
-    641,
-    641,
-    642,
-    642,
-    642,
-    642,
-    642,
-    642,
-    642,
-    642,
-    642,
-    642,
-    642,
-    641,
-    641,
-    641,
-    640,
-    639,
-    639,
-    638,
-    637,
-    637,
-    636,
-    635,
-    634,
-    633,
-    632,
-    631,
-    629,
-    628,
-    627,
-    625,
-    624,
-    622,
-    621,
-    619,
-    618,
-    616,
-    614,
-    612,
-    611,
-    609,
-    607,
-    605,
-    603,
-    601,
-    598,
-    596,
-    594,
-    592,
-    589,
-    587,
-    584,
-    582,
-    579,
-    577,
-    574,
-    571,
-    569,
-    566,
-    563,
-    560,
-    557,
-    555,
-    552,
-    549,
-    546,
-    542,
-    539,
-    536,
-    533,
-    530,
-    527,
-    523,
-    520,
-    517,
-    513,
-    510,
-    507,
-    503,
-    500,
-    496,
-    493,
-    489,
-    486,
-    482,
-    479,
-    475,
-    471,
-    468,
-    464,
-    460,
-    457,
-    453,
-    449,
-    445,
-    442,
-    438,
-    434,
-    430,
-    427,
-    423,
-    419,
-    415,
-    411,
-    408,
-    404,
-    400,
-    396,
-    392,
-    389,
-    385,
-    381,
-    377,
-    373,
-    370,
-    366,
-    362,
-    358,
-    355,
-    351,
-    347,
-    343,
-    340,
-    336,
-    332,
-    329,
-    325,
-    321,
-    318,
-    314,
-    311,
-    307,
-    304,
-    300,
-    297,
-    293,
-    290,
-    287,
-    283,
-    280,
-    277,
-    273,
-    270,
-    267,
-    264,
-    261,
-    258,
-    254,
-    251,
-    248,
-    245,
-    243,
-    240,
-    237,
-    234,
-    231,
-    229,
-    226,
-    223,
-    221,
-    218,
-    216,
-    213,
-    211,
-    208,
-    206,
-    204,
-    202,
-    199,
-    197,
-    195,
-    193,
-    191,
-    189,
-    188,
-    186,
-    184,
-    182,
-    181,
-    179,
-    178,
-    176,
-    175,
-    173,
-    172,
-    171,
-    169,
-    168,
-    167,
-    166,
-    165,
-    164,
-    163,
-    163,
-    162,
-    161,
-    161,
-    160,
-    159,
-    159,
-    159,
-    158,
-    158,
-    158,
-    158,
-    158,
-    158,
-    158,
-    158,
-    158,
-    158,
-    158,
-    159,
-    159,
-    159,
-    160,
-    161,
-    161,
-    162,
-    163,
-    163,
-    164,
-    165,
-    166,
-    167,
-    168,
-    169,
-    171,
-    172,
-    173,
-    175,
-    176,
-    178,
-    179,
-    181,
-    182,
-    184,
-    186,
-    188,
-    189,
-    191,
-    193,
-    195,
-    197,
-    199,
-    202,
-    204,
-    206,
-    208,
-    211,
-    213,
-    216,
-    218,
-    221,
-    223,
-    226,
-    229,
-    231,
-    234,
-    237,
-    240,
-    243,
-    245,
-    248,
-    251,
-    254,
-    258,
-    261,
-    264,
-    267,
-    270,
-    273,
-    277,
-    280,
-    283,
-    287,
-    290,
-    293,
-    297,
-    300,
-    304,
-    307,
-    311,
-    314,
-    318,
-    321,
-    325,
-    329,
-    332,
-    336,
-    340,
-    343,
-    347,
-    351,
-    355,
-    358,
-    362,
-    366,
-    370,
-    373,
-    377,
-    381,
-    385,
-    389,
-    392,
-    396,
-};
-
-#define LENGTH_500_SINE_TABLE_LENGTH 500
-uint32_t LENGTH_500_SINE_TABLE[] = {
-    500,
-    504,
-    508,
-    511,
-    515,
-    519,
-    523,
-    527,
-    530,
-    534,
-    538,
-    542,
-    546,
-    549,
-    553,
-    557,
-    561,
-    564,
-    568,
-    572,
-    575,
-    579,
-    583,
-    586,
-    590,
-    594,
-    597,
-    601,
-    604,
-    608,
-    612,
-    615,
-    619,
-    622,
-    626,
-    629,
-    632,
-    636,
-    639,
-    643,
-    646,
-    649,
-    653,
-    656,
-    659,
-    662,
-    666,
-    669,
-    672,
-    675,
-    678,
-    681,
-    684,
-    687,
-    690,
-    693,
-    696,
-    699,
-    702,
-    705,
-    707,
-    710,
-    713,
-    716,
-    718,
-    721,
-    723,
-    726,
-    729,
-    731,
-    733,
-    736,
-    738,
-    741,
-    743,
-    745,
-    747,
-    750,
-    752,
-    754,
-    756,
-    758,
-    760,
-    762,
-    764,
-    766,
-    767,
-    769,
-    771,
-    773,
-    774,
-    776,
-    777,
-    779,
-    780,
-    782,
-    783,
-    784,
-    786,
-    787,
-    788,
-    789,
-    790,
-    792,
-    793,
-    794,
-    794,
-    795,
-    796,
-    797,
-    798,
-    798,
-    799,
-    800,
-    800,
-    801,
-    801,
-    802,
-    802,
-    802,
-    802,
-    803,
-    803,
-    803,
-    803,
-    803,
-    803,
-    803,
-    803,
-    803,
-    802,
-    802,
-    802,
-    802,
-    801,
-    801,
-    800,
-    800,
-    799,
-    798,
-    798,
-    797,
-    796,
-    795,
-    794,
-    794,
-    793,
-    792,
-    790,
-    789,
-    788,
-    787,
-    786,
-    784,
-    783,
-    782,
-    780,
-    779,
-    777,
-    776,
-    774,
-    773,
-    771,
-    769,
-    767,
-    766,
-    764,
-    762,
-    760,
-    758,
-    756,
-    754,
-    752,
-    750,
-    747,
-    745,
-    743,
-    741,
-    738,
-    736,
-    733,
-    731,
-    729,
-    726,
-    723,
-    721,
-    718,
-    716,
-    713,
-    710,
-    707,
-    705,
-    702,
-    699,
-    696,
-    693,
-    690,
-    687,
-    684,
-    681,
-    678,
-    675,
-    672,
-    669,
-    666,
-    662,
-    659,
-    656,
-    653,
-    649,
-    646,
-    643,
-    639,
-    636,
-    632,
-    629,
-    626,
-    622,
-    619,
-    615,
-    612,
-    608,
-    604,
-    601,
-    597,
-    594,
-    590,
-    586,
-    583,
-    579,
-    575,
-    572,
-    568,
-    564,
-    561,
-    557,
-    553,
-    549,
-    546,
-    542,
-    538,
-    534,
-    530,
-    527,
-    523,
-    519,
-    515,
-    511,
-    508,
-    504,
-    500,
-    496,
-    492,
-    489,
-    485,
-    481,
-    477,
-    473,
-    470,
-    466,
-    462,
-    458,
-    454,
-    451,
-    447,
-    443,
-    439,
-    436,
-    432,
-    428,
-    425,
-    421,
-    417,
-    414,
-    410,
-    406,
-    403,
-    399,
-    396,
-    392,
-    388,
-    385,
-    381,
-    378,
-    374,
-    371,
-    368,
-    364,
-    361,
-    357,
-    354,
-    351,
-    347,
-    344,
-    341,
-    338,
-    334,
-    331,
-    328,
-    325,
-    322,
-    319,
-    316,
-    313,
-    310,
-    307,
-    304,
-    301,
-    298,
-    295,
-    293,
-    290,
-    287,
-    284,
-    282,
-    279,
-    277,
-    274,
-    271,
-    269,
-    267,
-    264,
-    262,
-    259,
-    257,
-    255,
-    253,
-    250,
-    248,
-    246,
-    244,
-    242,
-    240,
-    238,
-    236,
-    234,
-    233,
-    231,
-    229,
-    227,
-    226,
-    224,
-    223,
-    221,
-    220,
-    218,
-    217,
-    216,
-    214,
-    213,
-    212,
-    211,
-    210,
-    208,
-    207,
-    206,
-    206,
-    205,
-    204,
-    203,
-    202,
-    202,
-    201,
-    200,
-    200,
-    199,
-    199,
-    198,
-    198,
-    198,
-    198,
-    197,
-    197,
-    197,
-    197,
-    197,
-    197,
-    197,
-    197,
-    197,
-    198,
-    198,
-    198,
-    198,
-    199,
-    199,
-    200,
-    200,
-    201,
-    202,
-    202,
-    203,
-    204,
-    205,
-    206,
-    206,
-    207,
-    208,
-    210,
-    211,
-    212,
-    213,
-    214,
-    216,
-    217,
-    218,
-    220,
-    221,
-    223,
-    224,
-    226,
-    227,
-    229,
-    231,
-    233,
-    234,
-    236,
-    238,
-    240,
-    242,
-    244,
-    246,
-    248,
-    250,
-    253,
-    255,
-    257,
-    259,
-    262,
-    264,
-    267,
-    269,
-    271,
-    274,
-    277,
-    279,
-    282,
-    284,
-    287,
-    290,
-    293,
-    295,
-    298,
-    301,
-    304,
-    307,
-    310,
-    313,
-    316,
-    319,
-    322,
-    325,
-    328,
-    331,
-    334,
-    338,
-    341,
-    344,
-    347,
-    351,
-    354,
-    357,
-    361,
-    364,
-    368,
-    371,
-    374,
-    378,
-    381,
-    385,
-    388,
-    392,
-    396,
-    399,
-    403,
-    406,
-    410,
-    414,
-    417,
-    421,
-    425,
-    428,
-    432,
-    436,
-    439,
-    443,
-    447,
-    451,
-    454,
-    458,
-    462,
-    466,
-    470,
-    473,
-    477,
-    481,
-    485,
-    489,
-    492,
-    496,
-};
-
-#define LENGTH_600_SINE_TABLE_LENGTH 600
-uint32_t LENGTH_600_SINE_TABLE[] = {
-    600,
-    604,
-    608,
-    611,
-    615,
-    619,
-    623,
-    627,
-    630,
-    634,
-    638,
-    642,
-    646,
-    649,
-    653,
-    657,
-    661,
-    664,
-    668,
-    672,
-    676,
-    679,
-    683,
-    687,
-    690,
-    694,
-    698,
-    701,
-    705,
-    709,
-    712,
-    716,
-    720,
-    723,
-    727,
-    730,
-    734,
-    737,
-    741,
-    744,
-    748,
-    751,
-    755,
-    758,
-    762,
-    765,
-    768,
-    772,
-    775,
-    779,
-    782,
-    785,
-    788,
-    792,
-    795,
-    798,
-    801,
-    804,
-    808,
-    811,
-    814,
-    817,
-    820,
-    823,
-    826,
-    829,
-    832,
-    835,
-    838,
-    840,
-    843,
-    846,
-    849,
-    852,
-    854,
-    857,
-    860,
-    862,
-    865,
-    868,
-    870,
-    873,
-    875,
-    878,
-    880,
-    883,
-    885,
-    887,
-    890,
-    892,
-    894,
-    896,
-    899,
-    901,
-    903,
-    905,
-    907,
-    909,
-    911,
-    913,
-    915,
-    917,
-    919,
-    920,
-    922,
-    924,
-    926,
-    927,
-    929,
-    931,
-    932,
-    934,
-    935,
-    937,
-    938,
-    939,
-    941,
-    942,
-    943,
-    945,
-    946,
-    947,
-    948,
-    949,
-    950,
-    951,
-    952,
-    953,
-    954,
-    955,
-    956,
-    956,
-    957,
-    958,
-    959,
-    959,
-    960,
-    960,
-    961,
-    961,
-    962,
-    962,
-    962,
-    963,
-    963,
-    963,
-    963,
-    963,
-    964,
-    964,
-    964,
-    964,
-    964,
-    963,
-    963,
-    963,
-    963,
-    963,
-    962,
-    962,
-    962,
-    961,
-    961,
-    960,
-    960,
-    959,
-    959,
-    958,
-    957,
-    956,
-    956,
-    955,
-    954,
-    953,
-    952,
-    951,
-    950,
-    949,
-    948,
-    947,
-    946,
-    945,
-    943,
-    942,
-    941,
-    939,
-    938,
-    937,
-    935,
-    934,
-    932,
-    931,
-    929,
-    927,
-    926,
-    924,
-    922,
-    920,
-    919,
-    917,
-    915,
-    913,
-    911,
-    909,
-    907,
-    905,
-    903,
-    901,
-    899,
-    896,
-    894,
-    892,
-    890,
-    887,
-    885,
-    883,
-    880,
-    878,
-    875,
-    873,
-    870,
-    868,
-    865,
-    862,
-    860,
-    857,
-    854,
-    852,
-    849,
-    846,
-    843,
-    840,
-    838,
-    835,
-    832,
-    829,
-    826,
-    823,
-    820,
-    817,
-    814,
-    811,
-    808,
-    804,
-    801,
-    798,
-    795,
-    792,
-    788,
-    785,
-    782,
-    779,
-    775,
-    772,
-    768,
-    765,
-    762,
-    758,
-    755,
-    751,
-    748,
-    744,
-    741,
-    737,
-    734,
-    730,
-    727,
-    723,
-    720,
-    716,
-    712,
-    709,
-    705,
-    701,
-    698,
-    694,
-    690,
-    687,
-    683,
-    679,
-    676,
-    672,
-    668,
-    664,
-    661,
-    657,
-    653,
-    649,
-    646,
-    642,
-    638,
-    634,
-    630,
-    627,
-    623,
-    619,
-    615,
-    611,
-    608,
-    604,
-    600,
-    596,
-    592,
-    589,
-    585,
-    581,
-    577,
-    573,
-    570,
-    566,
-    562,
-    558,
-    554,
-    551,
-    547,
-    543,
-    539,
-    536,
-    532,
-    528,
-    524,
-    521,
-    517,
-    513,
-    510,
-    506,
-    502,
-    499,
-    495,
-    491,
-    488,
-    484,
-    480,
-    477,
-    473,
-    470,
-    466,
-    463,
-    459,
-    456,
-    452,
-    449,
-    445,
-    442,
-    438,
-    435,
-    432,
-    428,
-    425,
-    421,
-    418,
-    415,
-    412,
-    408,
-    405,
-    402,
-    399,
-    396,
-    392,
-    389,
-    386,
-    383,
-    380,
-    377,
-    374,
-    371,
-    368,
-    365,
-    362,
-    360,
-    357,
-    354,
-    351,
-    348,
-    346,
-    343,
-    340,
-    338,
-    335,
-    332,
-    330,
-    327,
-    325,
-    322,
-    320,
-    317,
-    315,
-    313,
-    310,
-    308,
-    306,
-    304,
-    301,
-    299,
-    297,
-    295,
-    293,
-    291,
-    289,
-    287,
-    285,
-    283,
-    281,
-    280,
-    278,
-    276,
-    274,
-    273,
-    271,
-    269,
-    268,
-    266,
-    265,
-    263,
-    262,
-    261,
-    259,
-    258,
-    257,
-    255,
-    254,
-    253,
-    252,
-    251,
-    250,
-    249,
-    248,
-    247,
-    246,
-    245,
-    244,
-    244,
-    243,
-    242,
-    241,
-    241,
-    240,
-    240,
-    239,
-    239,
-    238,
-    238,
-    238,
-    237,
-    237,
-    237,
-    237,
-    237,
-    236,
-    236,
-    236,
-    236,
-    236,
-    237,
-    237,
-    237,
-    237,
-    237,
-    238,
-    238,
-    238,
-    239,
-    239,
-    240,
-    240,
-    241,
-    241,
-    242,
-    243,
-    244,
-    244,
-    245,
-    246,
-    247,
-    248,
-    249,
-    250,
-    251,
-    252,
-    253,
-    254,
-    255,
-    257,
-    258,
-    259,
-    261,
-    262,
-    263,
-    265,
-    266,
-    268,
-    269,
-    271,
-    273,
-    274,
-    276,
-    278,
-    280,
-    281,
-    283,
-    285,
-    287,
-    289,
-    291,
-    293,
-    295,
-    297,
-    299,
-    301,
-    304,
-    306,
-    308,
-    310,
-    313,
-    315,
-    317,
-    320,
-    322,
-    325,
-    327,
-    330,
-    332,
-    335,
-    338,
-    340,
-    343,
-    346,
-    348,
-    351,
-    354,
-    357,
-    360,
-    362,
-    365,
-    368,
-    371,
-    374,
-    377,
-    380,
-    383,
-    386,
-    389,
-    392,
-    396,
-    399,
-    402,
-    405,
-    408,
-    412,
-    415,
-    418,
-    421,
-    425,
-    428,
-    432,
-    435,
-    438,
-    442,
-    445,
-    449,
-    452,
-    456,
-    459,
-    463,
-    466,
-    470,
-    473,
-    477,
-    480,
-    484,
-    488,
-    491,
-    495,
-    499,
-    502,
-    506,
-    510,
-    513,
-    517,
-    521,
-    524,
-    528,
-    532,
-    536,
-    539,
-    543,
-    547,
-    551,
-    554,
-    558,
-    562,
-    566,
-    570,
-    573,
-    577,
-    581,
-    585,
-    589,
-    592,
-    596,
-};
-
-#define LENGTH_700_SINE_TABLE_LENGTH 700
-uint32_t LENGTH_700_SINE_TABLE[] = {
-    700,
-    704,
-    708,
-    711,
-    715,
-    719,
-    723,
-    727,
-    730,
-    734,
-    738,
-    742,
-    746,
-    749,
-    753,
-    757,
-    761,
-    764,
-    768,
-    772,
-    776,
-    779,
-    783,
-    787,
-    791,
-    794,
-    798,
-    802,
-    806,
-    809,
-    813,
-    817,
-    820,
-    824,
-    827,
-    831,
-    835,
-    838,
-    842,
-    845,
-    849,
-    853,
-    856,
-    860,
-    863,
-    867,
-    870,
-    874,
-    877,
-    881,
-    884,
-    887,
-    891,
-    894,
-    898,
-    901,
-    904,
-    908,
-    911,
-    914,
-    918,
-    921,
-    924,
-    927,
-    931,
-    934,
-    937,
-    940,
-    943,
-    946,
-    949,
-    952,
-    955,
-    959,
-    962,
-    965,
-    967,
-    970,
-    973,
-    976,
-    979,
-    982,
-    985,
-    988,
-    990,
-    993,
-    996,
-    999,
-    1001,
-    1004,
-    1007,
-    1009,
-    1012,
-    1014,
-    1017,
-    1019,
-    1022,
-    1024,
-    1027,
-    1029,
-    1032,
-    1034,
-    1036,
-    1039,
-    1041,
-    1043,
-    1045,
-    1048,
-    1050,
-    1052,
-    1054,
-    1056,
-    1058,
-    1060,
-    1062,
-    1064,
-    1066,
-    1068,
-    1070,
-    1072,
-    1074,
-    1075,
-    1077,
-    1079,
-    1081,
-    1082,
-    1084,
-    1085,
-    1087,
-    1089,
-    1090,
-    1092,
-    1093,
-    1094,
-    1096,
-    1097,
-    1099,
-    1100,
-    1101,
-    1102,
-    1103,
-    1105,
-    1106,
-    1107,
-    1108,
-    1109,
-    1110,
-    1111,
-    1112,
-    1113,
-    1114,
-    1114,
-    1115,
-    1116,
-    1117,
-    1117,
-    1118,
-    1119,
-    1119,
-    1120,
-    1120,
-    1121,
-    1121,
-    1122,
-    1122,
-    1123,
-    1123,
-    1123,
-    1123,
-    1124,
-    1124,
-    1124,
-    1124,
-    1124,
-    1124,
-    1124,
-    1124,
-    1124,
-    1124,
-    1124,
-    1124,
-    1124,
-    1123,
-    1123,
-    1123,
-    1123,
-    1122,
-    1122,
-    1121,
-    1121,
-    1120,
-    1120,
-    1119,
-    1119,
-    1118,
-    1117,
-    1117,
-    1116,
-    1115,
-    1114,
-    1114,
-    1113,
-    1112,
-    1111,
-    1110,
-    1109,
-    1108,
-    1107,
-    1106,
-    1105,
-    1103,
-    1102,
-    1101,
-    1100,
-    1099,
-    1097,
-    1096,
-    1094,
-    1093,
-    1092,
-    1090,
-    1089,
-    1087,
-    1085,
-    1084,
-    1082,
-    1081,
-    1079,
-    1077,
-    1075,
-    1074,
-    1072,
-    1070,
-    1068,
-    1066,
-    1064,
-    1062,
-    1060,
-    1058,
-    1056,
-    1054,
-    1052,
-    1050,
-    1048,
-    1045,
-    1043,
-    1041,
-    1039,
-    1036,
-    1034,
-    1032,
-    1029,
-    1027,
-    1024,
-    1022,
-    1019,
-    1017,
-    1014,
-    1012,
-    1009,
-    1007,
-    1004,
-    1001,
-    999,
-    996,
-    993,
-    990,
-    988,
-    985,
-    982,
-    979,
-    976,
-    973,
-    970,
-    967,
-    965,
-    962,
-    959,
-    955,
-    952,
-    949,
-    946,
-    943,
-    940,
-    937,
-    934,
-    931,
-    927,
-    924,
-    921,
-    918,
-    914,
-    911,
-    908,
-    904,
-    901,
-    898,
-    894,
-    891,
-    887,
-    884,
-    881,
-    877,
-    874,
-    870,
-    867,
-    863,
-    860,
-    856,
-    853,
-    849,
-    845,
-    842,
-    838,
-    835,
-    831,
-    827,
-    824,
-    820,
-    817,
-    813,
-    809,
-    806,
-    802,
-    798,
-    794,
-    791,
-    787,
-    783,
-    779,
-    776,
-    772,
-    768,
-    764,
-    761,
-    757,
-    753,
-    749,
-    746,
-    742,
-    738,
-    734,
-    730,
-    727,
-    723,
-    719,
-    715,
-    711,
-    708,
-    704,
-    700,
-    696,
-    692,
-    689,
-    685,
-    681,
-    677,
-    673,
-    670,
-    666,
-    662,
-    658,
-    654,
-    651,
-    647,
-    643,
-    639,
-    636,
-    632,
-    628,
-    624,
-    621,
-    617,
-    613,
-    609,
-    606,
-    602,
-    598,
-    594,
-    591,
-    587,
-    583,
-    580,
-    576,
-    573,
-    569,
-    565,
-    562,
-    558,
-    555,
-    551,
-    547,
-    544,
-    540,
-    537,
-    533,
-    530,
-    526,
-    523,
-    519,
-    516,
-    513,
-    509,
-    506,
-    502,
-    499,
-    496,
-    492,
-    489,
-    486,
-    482,
-    479,
-    476,
-    473,
-    469,
-    466,
-    463,
-    460,
-    457,
-    454,
-    451,
-    448,
-    445,
-    441,
-    438,
-    435,
-    433,
-    430,
-    427,
-    424,
-    421,
-    418,
-    415,
-    412,
-    410,
-    407,
-    404,
-    401,
-    399,
-    396,
-    393,
-    391,
-    388,
-    386,
-    383,
-    381,
-    378,
-    376,
-    373,
-    371,
-    368,
-    366,
-    364,
-    361,
-    359,
-    357,
-    355,
-    352,
-    350,
-    348,
-    346,
-    344,
-    342,
-    340,
-    338,
-    336,
-    334,
-    332,
-    330,
-    328,
-    326,
-    325,
-    323,
-    321,
-    319,
-    318,
-    316,
-    315,
-    313,
-    311,
-    310,
-    308,
-    307,
-    306,
-    304,
-    303,
-    301,
-    300,
-    299,
-    298,
-    297,
-    295,
-    294,
-    293,
-    292,
-    291,
-    290,
-    289,
-    288,
-    287,
-    286,
-    286,
-    285,
-    284,
-    283,
-    283,
-    282,
-    281,
-    281,
-    280,
-    280,
-    279,
-    279,
-    278,
-    278,
-    277,
-    277,
-    277,
-    277,
-    276,
-    276,
-    276,
-    276,
-    276,
-    276,
-    276,
-    276,
-    276,
-    276,
-    276,
-    276,
-    276,
-    277,
-    277,
-    277,
-    277,
-    278,
-    278,
-    279,
-    279,
-    280,
-    280,
-    281,
-    281,
-    282,
-    283,
-    283,
-    284,
-    285,
-    286,
-    286,
-    287,
-    288,
-    289,
-    290,
-    291,
-    292,
-    293,
-    294,
-    295,
-    297,
-    298,
-    299,
-    300,
-    301,
-    303,
-    304,
-    306,
-    307,
-    308,
-    310,
-    311,
-    313,
-    315,
-    316,
-    318,
-    319,
-    321,
-    323,
-    325,
-    326,
-    328,
-    330,
-    332,
-    334,
-    336,
-    338,
-    340,
-    342,
-    344,
-    346,
-    348,
-    350,
-    352,
-    355,
-    357,
-    359,
-    361,
-    364,
-    366,
-    368,
-    371,
-    373,
-    376,
-    378,
-    381,
-    383,
-    386,
-    388,
-    391,
-    393,
-    396,
-    399,
-    401,
-    404,
-    407,
-    410,
-    412,
-    415,
-    418,
-    421,
-    424,
-    427,
-    430,
-    433,
-    435,
-    438,
-    441,
-    445,
-    448,
-    451,
-    454,
-    457,
-    460,
-    463,
-    466,
-    469,
-    473,
-    476,
-    479,
-    482,
-    486,
-    489,
-    492,
-    496,
-    499,
-    502,
-    506,
-    509,
-    513,
-    516,
-    519,
-    523,
-    526,
-    530,
-    533,
-    537,
-    540,
-    544,
-    547,
-    551,
-    555,
-    558,
-    562,
-    565,
-    569,
-    573,
-    576,
-    580,
-    583,
-    587,
-    591,
-    594,
-    598,
-    602,
-    606,
-    609,
-    613,
-    617,
-    621,
-    624,
-    628,
-    632,
-    636,
-    639,
-    643,
-    647,
-    651,
-    654,
-    658,
-    662,
-    666,
-    670,
-    673,
-    677,
-    681,
-    685,
-    689,
-    692,
-    696,
-};
-
-#define LENGTH_800_SINE_TABLE_LENGTH 800
-uint32_t LENGTH_800_SINE_TABLE[] = {
-    800,
-    804,
-    808,
-    811,
-    815,
-    819,
-    823,
-    827,
-    830,
-    834,
-    838,
-    842,
-    846,
-    849,
-    853,
-    857,
-    861,
-    865,
-    868,
-    872,
-    876,
-    880,
-    883,
-    887,
-    891,
-    895,
-    898,
-    902,
-    906,
-    909,
-    913,
-    917,
-    921,
-    924,
-    928,
-    932,
-    935,
-    939,
-    943,
-    946,
-    950,
-    953,
-    957,
-    961,
-    964,
-    968,
-    971,
-    975,
-    978,
-    982,
-    986,
-    989,
-    993,
-    996,
-    1000,
-    1003,
-    1006,
-    1010,
-    1013,
-    1017,
-    1020,
-    1024,
-    1027,
-    1030,
-    1034,
-    1037,
-    1040,
-    1044,
-    1047,
-    1050,
-    1053,
-    1057,
-    1060,
-    1063,
-    1066,
-    1069,
-    1073,
-    1076,
-    1079,
-    1082,
-    1085,
-    1088,
-    1091,
-    1094,
-    1097,
-    1100,
-    1103,
-    1106,
-    1109,
-    1112,
-    1115,
-    1118,
-    1121,
-    1123,
-    1126,
-    1129,
-    1132,
-    1135,
-    1137,
-    1140,
-    1143,
-    1146,
-    1148,
-    1151,
-    1153,
-    1156,
-    1159,
-    1161,
-    1164,
-    1166,
-    1169,
-    1171,
-    1174,
-    1176,
-    1178,
-    1181,
-    1183,
-    1185,
-    1188,
-    1190,
-    1192,
-    1194,
-    1197,
-    1199,
-    1201,
-    1203,
-    1205,
-    1207,
-    1209,
-    1211,
-    1213,
-    1215,
-    1217,
-    1219,
-    1221,
-    1223,
-    1225,
-    1227,
-    1228,
-    1230,
-    1232,
-    1234,
-    1235,
-    1237,
-    1239,
-    1240,
-    1242,
-    1243,
-    1245,
-    1246,
-    1248,
-    1249,
-    1251,
-    1252,
-    1254,
-    1255,
-    1256,
-    1257,
-    1259,
-    1260,
-    1261,
-    1262,
-    1263,
-    1265,
-    1266,
-    1267,
-    1268,
-    1269,
-    1270,
-    1271,
-    1271,
-    1272,
-    1273,
-    1274,
-    1275,
-    1276,
-    1276,
-    1277,
-    1278,
-    1278,
-    1279,
-    1279,
-    1280,
-    1281,
-    1281,
-    1281,
-    1282,
-    1282,
-    1283,
-    1283,
-    1283,
-    1284,
-    1284,
-    1284,
-    1284,
-    1284,
-    1285,
-    1285,
-    1285,
-    1285,
-    1285,
-    1285,
-    1285,
-    1285,
-    1285,
-    1284,
-    1284,
-    1284,
-    1284,
-    1284,
-    1283,
-    1283,
-    1283,
-    1282,
-    1282,
-    1281,
-    1281,
-    1281,
-    1280,
-    1279,
-    1279,
-    1278,
-    1278,
-    1277,
-    1276,
-    1276,
-    1275,
-    1274,
-    1273,
-    1272,
-    1271,
-    1271,
-    1270,
-    1269,
-    1268,
-    1267,
-    1266,
-    1265,
-    1263,
-    1262,
-    1261,
-    1260,
-    1259,
-    1257,
-    1256,
-    1255,
-    1254,
-    1252,
-    1251,
-    1249,
-    1248,
-    1246,
-    1245,
-    1243,
-    1242,
-    1240,
-    1239,
-    1237,
-    1235,
-    1234,
-    1232,
-    1230,
-    1228,
-    1227,
-    1225,
-    1223,
-    1221,
-    1219,
-    1217,
-    1215,
-    1213,
-    1211,
-    1209,
-    1207,
-    1205,
-    1203,
-    1201,
-    1199,
-    1197,
-    1194,
-    1192,
-    1190,
-    1188,
-    1185,
-    1183,
-    1181,
-    1178,
-    1176,
-    1174,
-    1171,
-    1169,
-    1166,
-    1164,
-    1161,
-    1159,
-    1156,
-    1153,
-    1151,
-    1148,
-    1146,
-    1143,
-    1140,
-    1137,
-    1135,
-    1132,
-    1129,
-    1126,
-    1123,
-    1121,
-    1118,
-    1115,
-    1112,
-    1109,
-    1106,
-    1103,
-    1100,
-    1097,
-    1094,
-    1091,
-    1088,
-    1085,
-    1082,
-    1079,
-    1076,
-    1073,
-    1069,
-    1066,
-    1063,
-    1060,
-    1057,
-    1053,
-    1050,
-    1047,
-    1044,
-    1040,
-    1037,
-    1034,
-    1030,
-    1027,
-    1024,
-    1020,
-    1017,
-    1013,
-    1010,
-    1006,
-    1003,
-    1000,
-    996,
-    993,
-    989,
-    986,
-    982,
-    978,
-    975,
-    971,
-    968,
-    964,
-    961,
-    957,
-    953,
-    950,
-    946,
-    943,
-    939,
-    935,
-    932,
-    928,
-    924,
-    921,
-    917,
-    913,
-    909,
-    906,
-    902,
-    898,
-    895,
-    891,
-    887,
-    883,
-    880,
-    876,
-    872,
-    868,
-    865,
-    861,
-    857,
-    853,
-    849,
-    846,
-    842,
-    838,
-    834,
-    830,
-    827,
-    823,
-    819,
-    815,
-    811,
-    808,
-    804,
-    800,
-    796,
-    792,
-    789,
-    785,
-    781,
-    777,
-    773,
-    770,
-    766,
-    762,
-    758,
-    754,
-    751,
-    747,
-    743,
-    739,
-    735,
-    732,
-    728,
-    724,
-    720,
-    717,
-    713,
-    709,
-    705,
-    702,
-    698,
-    694,
-    691,
-    687,
-    683,
-    679,
-    676,
-    672,
-    668,
-    665,
-    661,
-    657,
-    654,
-    650,
-    647,
-    643,
-    639,
-    636,
-    632,
-    629,
-    625,
-    622,
-    618,
-    614,
-    611,
-    607,
-    604,
-    600,
-    597,
-    594,
-    590,
-    587,
-    583,
-    580,
-    576,
-    573,
-    570,
-    566,
-    563,
-    560,
-    556,
-    553,
-    550,
-    547,
-    543,
-    540,
-    537,
-    534,
-    531,
-    527,
-    524,
-    521,
-    518,
-    515,
-    512,
-    509,
-    506,
-    503,
-    500,
-    497,
-    494,
-    491,
-    488,
-    485,
-    482,
-    479,
-    477,
-    474,
-    471,
-    468,
-    465,
-    463,
-    460,
-    457,
-    454,
-    452,
-    449,
-    447,
-    444,
-    441,
-    439,
-    436,
-    434,
-    431,
-    429,
-    426,
-    424,
-    422,
-    419,
-    417,
-    415,
-    412,
-    410,
-    408,
-    406,
-    403,
-    401,
-    399,
-    397,
-    395,
-    393,
-    391,
-    389,
-    387,
-    385,
-    383,
-    381,
-    379,
-    377,
-    375,
-    373,
-    372,
-    370,
-    368,
-    366,
-    365,
-    363,
-    361,
-    360,
-    358,
-    357,
-    355,
-    354,
-    352,
-    351,
-    349,
-    348,
-    346,
-    345,
-    344,
-    343,
-    341,
-    340,
-    339,
-    338,
-    337,
-    335,
-    334,
-    333,
-    332,
-    331,
-    330,
-    329,
-    329,
-    328,
-    327,
-    326,
-    325,
-    324,
-    324,
-    323,
-    322,
-    322,
-    321,
-    321,
-    320,
-    319,
-    319,
-    319,
-    318,
-    318,
-    317,
-    317,
-    317,
-    316,
-    316,
-    316,
-    316,
-    316,
-    315,
-    315,
-    315,
-    315,
-    315,
-    315,
-    315,
-    315,
-    315,
-    316,
-    316,
-    316,
-    316,
-    316,
-    317,
-    317,
-    317,
-    318,
-    318,
-    319,
-    319,
-    319,
-    320,
-    321,
-    321,
-    322,
-    322,
-    323,
-    324,
-    324,
-    325,
-    326,
-    327,
-    328,
-    329,
-    329,
-    330,
-    331,
-    332,
-    333,
-    334,
-    335,
-    337,
-    338,
-    339,
-    340,
-    341,
-    343,
-    344,
-    345,
-    346,
-    348,
-    349,
-    351,
-    352,
-    354,
-    355,
-    357,
-    358,
-    360,
-    361,
-    363,
-    365,
-    366,
-    368,
-    370,
-    372,
-    373,
-    375,
-    377,
-    379,
-    381,
-    383,
-    385,
-    387,
-    389,
-    391,
-    393,
-    395,
-    397,
-    399,
-    401,
-    403,
-    406,
-    408,
-    410,
-    412,
-    415,
-    417,
-    419,
-    422,
-    424,
-    426,
-    429,
-    431,
-    434,
-    436,
-    439,
-    441,
-    444,
-    447,
-    449,
-    452,
-    454,
-    457,
-    460,
-    463,
-    465,
-    468,
-    471,
-    474,
-    477,
-    479,
-    482,
-    485,
-    488,
-    491,
-    494,
-    497,
-    500,
-    503,
-    506,
-    509,
-    512,
-    515,
-    518,
-    521,
-    524,
-    527,
-    531,
-    534,
-    537,
-    540,
-    543,
-    547,
-    550,
-    553,
-    556,
-    560,
-    563,
-    566,
-    570,
-    573,
-    576,
-    580,
-    583,
-    587,
-    590,
-    594,
-    597,
-    600,
-    604,
-    607,
-    611,
-    614,
-    618,
-    622,
-    625,
-    629,
-    632,
-    636,
-    639,
-    643,
-    647,
-    650,
-    654,
-    657,
-    661,
-    665,
-    668,
-    672,
-    676,
-    679,
-    683,
-    687,
-    691,
-    694,
-    698,
-    702,
-    705,
-    709,
-    713,
-    717,
-    720,
-    724,
-    728,
-    732,
-    735,
-    739,
-    743,
-    747,
-    751,
-    754,
-    758,
-    762,
-    766,
-    770,
-    773,
-    777,
-    781,
-    785,
-    789,
-    792,
-    796,
-};
-
-#define LENGTH_900_SINE_TABLE_LENGTH 900
-uint32_t LENGTH_900_SINE_TABLE[] = {
-    900,
-    904,
-    908,
-    911,
-    915,
-    919,
-    923,
-    927,
-    930,
-    934,
-    938,
-    942,
-    946,
-    949,
-    953,
-    957,
-    961,
-    965,
-    968,
-    972,
-    976,
-    980,
-    983,
-    987,
-    991,
-    995,
-    998,
-    1002,
-    1006,
-    1010,
-    1013,
-    1017,
-    1021,
-    1025,
-    1028,
-    1032,
-    1036,
-    1039,
-    1043,
-    1047,
-    1050,
-    1054,
-    1058,
-    1061,
-    1065,
-    1069,
-    1072,
-    1076,
-    1079,
-    1083,
-    1087,
-    1090,
-    1094,
-    1097,
-    1101,
-    1104,
-    1108,
-    1111,
-    1115,
-    1118,
-    1122,
-    1125,
-    1129,
-    1132,
-    1136,
-    1139,
-    1143,
-    1146,
-    1149,
-    1153,
-    1156,
-    1159,
-    1163,
-    1166,
-    1169,
-    1173,
-    1176,
-    1179,
-    1183,
-    1186,
-    1189,
-    1192,
-    1195,
-    1199,
-    1202,
-    1205,
-    1208,
-    1211,
-    1214,
-    1218,
-    1221,
-    1224,
-    1227,
-    1230,
-    1233,
-    1236,
-    1239,
-    1242,
-    1245,
-    1248,
-    1251,
-    1254,
-    1256,
-    1259,
-    1262,
-    1265,
-    1268,
-    1271,
-    1273,
-    1276,
-    1279,
-    1282,
-    1284,
-    1287,
-    1290,
-    1292,
-    1295,
-    1298,
-    1300,
-    1303,
-    1305,
-    1308,
-    1310,
-    1313,
-    1315,
-    1318,
-    1320,
-    1323,
-    1325,
-    1327,
-    1330,
-    1332,
-    1334,
-    1337,
-    1339,
-    1341,
-    1344,
-    1346,
-    1348,
-    1350,
-    1352,
-    1354,
-    1356,
-    1358,
-    1361,
-    1363,
-    1365,
-    1367,
-    1369,
-    1370,
-    1372,
-    1374,
-    1376,
-    1378,
-    1380,
-    1382,
-    1383,
-    1385,
-    1387,
-    1389,
-    1390,
-    1392,
-    1394,
-    1395,
-    1397,
-    1398,
-    1400,
-    1401,
-    1403,
-    1404,
-    1406,
-    1407,
-    1409,
-    1410,
-    1411,
-    1413,
-    1414,
-    1415,
-    1416,
-    1418,
-    1419,
-    1420,
-    1421,
-    1422,
-    1423,
-    1424,
-    1425,
-    1426,
-    1427,
-    1428,
-    1429,
-    1430,
-    1431,
-    1432,
-    1433,
-    1434,
-    1434,
-    1435,
-    1436,
-    1436,
-    1437,
-    1438,
-    1438,
-    1439,
-    1440,
-    1440,
-    1441,
-    1441,
-    1442,
-    1442,
-    1442,
-    1443,
-    1443,
-    1444,
-    1444,
-    1444,
-    1444,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1445,
-    1444,
-    1444,
-    1444,
-    1444,
-    1443,
-    1443,
-    1442,
-    1442,
-    1442,
-    1441,
-    1441,
-    1440,
-    1440,
-    1439,
-    1438,
-    1438,
-    1437,
-    1436,
-    1436,
-    1435,
-    1434,
-    1434,
-    1433,
-    1432,
-    1431,
-    1430,
-    1429,
-    1428,
-    1427,
-    1426,
-    1425,
-    1424,
-    1423,
-    1422,
-    1421,
-    1420,
-    1419,
-    1418,
-    1416,
-    1415,
-    1414,
-    1413,
-    1411,
-    1410,
-    1409,
-    1407,
-    1406,
-    1404,
-    1403,
-    1401,
-    1400,
-    1398,
-    1397,
-    1395,
-    1394,
-    1392,
-    1390,
-    1389,
-    1387,
-    1385,
-    1383,
-    1382,
-    1380,
-    1378,
-    1376,
-    1374,
-    1372,
-    1370,
-    1369,
-    1367,
-    1365,
-    1363,
-    1361,
-    1358,
-    1356,
-    1354,
-    1352,
-    1350,
-    1348,
-    1346,
-    1344,
-    1341,
-    1339,
-    1337,
-    1334,
-    1332,
-    1330,
-    1327,
-    1325,
-    1323,
-    1320,
-    1318,
-    1315,
-    1313,
-    1310,
-    1308,
-    1305,
-    1303,
-    1300,
-    1298,
-    1295,
-    1292,
-    1290,
-    1287,
-    1284,
-    1282,
-    1279,
-    1276,
-    1273,
-    1271,
-    1268,
-    1265,
-    1262,
-    1259,
-    1256,
-    1254,
-    1251,
-    1248,
-    1245,
-    1242,
-    1239,
-    1236,
-    1233,
-    1230,
-    1227,
-    1224,
-    1221,
-    1218,
-    1214,
-    1211,
-    1208,
-    1205,
-    1202,
-    1199,
-    1195,
-    1192,
-    1189,
-    1186,
-    1183,
-    1179,
-    1176,
-    1173,
-    1169,
-    1166,
-    1163,
-    1159,
-    1156,
-    1153,
-    1149,
-    1146,
-    1143,
-    1139,
-    1136,
-    1132,
-    1129,
-    1125,
-    1122,
-    1118,
-    1115,
-    1111,
-    1108,
-    1104,
-    1101,
-    1097,
-    1094,
-    1090,
-    1087,
-    1083,
-    1079,
-    1076,
-    1072,
-    1069,
-    1065,
-    1061,
-    1058,
-    1054,
-    1050,
-    1047,
-    1043,
-    1039,
-    1036,
-    1032,
-    1028,
-    1025,
-    1021,
-    1017,
-    1013,
-    1010,
-    1006,
-    1002,
-    998,
-    995,
-    991,
-    987,
-    983,
-    980,
-    976,
-    972,
-    968,
-    965,
-    961,
-    957,
-    953,
-    949,
-    946,
-    942,
-    938,
-    934,
-    930,
-    927,
-    923,
-    919,
-    915,
-    911,
-    908,
-    904,
-    900,
-    896,
-    892,
-    889,
-    885,
-    881,
-    877,
-    873,
-    870,
-    866,
-    862,
-    858,
-    854,
-    851,
-    847,
-    843,
-    839,
-    835,
-    832,
-    828,
-    824,
-    820,
-    817,
-    813,
-    809,
-    805,
-    802,
-    798,
-    794,
-    790,
-    787,
-    783,
-    779,
-    775,
-    772,
-    768,
-    764,
-    761,
-    757,
-    753,
-    750,
-    746,
-    742,
-    739,
-    735,
-    731,
-    728,
-    724,
-    721,
-    717,
-    713,
-    710,
-    706,
-    703,
-    699,
-    696,
-    692,
-    689,
-    685,
-    682,
-    678,
-    675,
-    671,
-    668,
-    664,
-    661,
-    657,
-    654,
-    651,
-    647,
-    644,
-    641,
-    637,
-    634,
-    631,
-    627,
-    624,
-    621,
-    617,
-    614,
-    611,
-    608,
-    605,
-    601,
-    598,
-    595,
-    592,
-    589,
-    586,
-    582,
-    579,
-    576,
-    573,
-    570,
-    567,
-    564,
-    561,
-    558,
-    555,
-    552,
-    549,
-    546,
-    544,
-    541,
-    538,
-    535,
-    532,
-    529,
-    527,
-    524,
-    521,
-    518,
-    516,
-    513,
-    510,
-    508,
-    505,
-    502,
-    500,
-    497,
-    495,
-    492,
-    490,
-    487,
-    485,
-    482,
-    480,
-    477,
-    475,
-    473,
-    470,
-    468,
-    466,
-    463,
-    461,
-    459,
-    456,
-    454,
-    452,
-    450,
-    448,
-    446,
-    444,
-    442,
-    439,
-    437,
-    435,
-    433,
-    431,
-    430,
-    428,
-    426,
-    424,
-    422,
-    420,
-    418,
-    417,
-    415,
-    413,
-    411,
-    410,
-    408,
-    406,
-    405,
-    403,
-    402,
-    400,
-    399,
-    397,
-    396,
-    394,
-    393,
-    391,
-    390,
-    389,
-    387,
-    386,
-    385,
-    384,
-    382,
-    381,
-    380,
-    379,
-    378,
-    377,
-    376,
-    375,
-    374,
-    373,
-    372,
-    371,
-    370,
-    369,
-    368,
-    367,
-    366,
-    366,
-    365,
-    364,
-    364,
-    363,
-    362,
-    362,
-    361,
-    360,
-    360,
-    359,
-    359,
-    358,
-    358,
-    358,
-    357,
-    357,
-    356,
-    356,
-    356,
-    356,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    355,
-    356,
-    356,
-    356,
-    356,
-    357,
-    357,
-    358,
-    358,
-    358,
-    359,
-    359,
-    360,
-    360,
-    361,
-    362,
-    362,
-    363,
-    364,
-    364,
-    365,
-    366,
-    366,
-    367,
-    368,
-    369,
-    370,
-    371,
-    372,
-    373,
-    374,
-    375,
-    376,
-    377,
-    378,
-    379,
-    380,
-    381,
-    382,
-    384,
-    385,
-    386,
-    387,
-    389,
-    390,
-    391,
-    393,
-    394,
-    396,
-    397,
-    399,
-    400,
-    402,
-    403,
-    405,
-    406,
-    408,
-    410,
-    411,
-    413,
-    415,
-    417,
-    418,
-    420,
-    422,
-    424,
-    426,
-    428,
-    430,
-    431,
-    433,
-    435,
-    437,
-    439,
-    442,
-    444,
-    446,
-    448,
-    450,
-    452,
-    454,
-    456,
-    459,
-    461,
-    463,
-    466,
-    468,
-    470,
-    473,
-    475,
-    477,
-    480,
-    482,
-    485,
-    487,
-    490,
-    492,
-    495,
-    497,
-    500,
-    502,
-    505,
-    508,
-    510,
-    513,
-    516,
-    518,
-    521,
-    524,
-    527,
-    529,
-    532,
-    535,
-    538,
-    541,
-    544,
-    546,
-    549,
-    552,
-    555,
-    558,
-    561,
-    564,
-    567,
-    570,
-    573,
-    576,
-    579,
-    582,
-    586,
-    589,
-    592,
-    595,
-    598,
-    601,
-    605,
-    608,
-    611,
-    614,
-    617,
-    621,
-    624,
-    627,
-    631,
-    634,
-    637,
-    641,
-    644,
-    647,
-    651,
-    654,
-    657,
-    661,
-    664,
-    668,
-    671,
-    675,
-    678,
-    682,
-    685,
-    689,
-    692,
-    696,
-    699,
-    703,
-    706,
-    710,
-    713,
-    717,
-    721,
-    724,
-    728,
-    731,
-    735,
-    739,
-    742,
-    746,
-    750,
-    753,
-    757,
-    761,
-    764,
-    768,
-    772,
-    775,
-    779,
-    783,
-    787,
-    790,
-    794,
-    798,
-    802,
-    805,
-    809,
-    813,
-    817,
-    820,
-    824,
-    828,
-    832,
-    835,
-    839,
-    843,
-    847,
-    851,
-    854,
-    858,
-    862,
-    866,
-    870,
-    873,
-    877,
-    881,
-    885,
-    889,
-    892,
-    896,
-};
-
-#define LENGTH_1000_SINE_TABLE_LENGTH 1000
-uint32_t LENGTH_1000_SINE_TABLE[] = {
-    1000,
-    1004,
-    1008,
-    1011,
-    1015,
-    1019,
-    1023,
-    1027,
-    1030,
-    1034,
-    1038,
-    1042,
-    1046,
-    1049,
-    1053,
-    1057,
-    1061,
-    1065,
-    1068,
-    1072,
-    1076,
-    1080,
-    1084,
-    1087,
-    1091,
-    1095,
-    1099,
-    1102,
-    1106,
-    1110,
-    1114,
-    1117,
-    1121,
-    1125,
-    1128,
-    1132,
-    1136,
-    1140,
-    1143,
-    1147,
-    1151,
-    1154,
-    1158,
-    1162,
-    1165,
-    1169,
-    1173,
-    1176,
-    1180,
-    1184,
-    1187,
-    1191,
-    1195,
-    1198,
-    1202,
-    1205,
-    1209,
-    1212,
-    1216,
-    1220,
-    1223,
-    1227,
-    1230,
-    1234,
-    1237,
-    1241,
-    1244,
-    1248,
-    1251,
-    1255,
-    1258,
-    1261,
-    1265,
-    1268,
-    1272,
-    1275,
-    1279,
-    1282,
-    1285,
-    1289,
-    1292,
-    1295,
-    1299,
-    1302,
-    1305,
-    1309,
-    1312,
-    1315,
-    1318,
-    1322,
-    1325,
-    1328,
-    1331,
-    1334,
-    1338,
-    1341,
-    1344,
-    1347,
-    1350,
-    1353,
-    1356,
-    1359,
-    1362,
-    1365,
-    1368,
-    1371,
-    1374,
-    1377,
-    1380,
-    1383,
-    1386,
-    1389,
-    1392,
-    1395,
-    1398,
-    1401,
-    1404,
-    1406,
-    1409,
-    1412,
-    1415,
-    1418,
-    1420,
-    1423,
-    1426,
-    1429,
-    1431,
-    1434,
-    1437,
-    1439,
-    1442,
-    1444,
-    1447,
-    1450,
-    1452,
-    1455,
-    1457,
-    1460,
-    1462,
-    1465,
-    1467,
-    1469,
-    1472,
-    1474,
-    1477,
-    1479,
-    1481,
-    1484,
-    1486,
-    1488,
-    1490,
-    1493,
-    1495,
-    1497,
-    1499,
-    1501,
-    1503,
-    1506,
-    1508,
-    1510,
-    1512,
-    1514,
-    1516,
-    1518,
-    1520,
-    1522,
-    1524,
-    1525,
-    1527,
-    1529,
-    1531,
-    1533,
-    1535,
-    1537,
-    1538,
-    1540,
-    1542,
-    1543,
-    1545,
-    1547,
-    1548,
-    1550,
-    1552,
-    1553,
-    1555,
-    1556,
-    1558,
-    1559,
-    1561,
-    1562,
-    1564,
-    1565,
-    1566,
-    1568,
-    1569,
-    1570,
-    1572,
-    1573,
-    1574,
-    1575,
-    1576,
-    1578,
-    1579,
-    1580,
-    1581,
-    1582,
-    1583,
-    1584,
-    1585,
-    1586,
-    1587,
-    1588,
-    1589,
-    1590,
-    1591,
-    1591,
-    1592,
-    1593,
-    1594,
-    1595,
-    1595,
-    1596,
-    1597,
-    1597,
-    1598,
-    1599,
-    1599,
-    1600,
-    1600,
-    1601,
-    1601,
-    1602,
-    1602,
-    1603,
-    1603,
-    1603,
-    1604,
-    1604,
-    1604,
-    1605,
-    1605,
-    1605,
-    1605,
-    1605,
-    1606,
-    1606,
-    1606,
-    1606,
-    1606,
-    1606,
-    1606,
-    1606,
-    1606,
-    1606,
-    1606,
-    1606,
-    1606,
-    1605,
-    1605,
-    1605,
-    1605,
-    1605,
-    1604,
-    1604,
-    1604,
-    1603,
-    1603,
-    1603,
-    1602,
-    1602,
-    1601,
-    1601,
-    1600,
-    1600,
-    1599,
-    1599,
-    1598,
-    1597,
-    1597,
-    1596,
-    1595,
-    1595,
-    1594,
-    1593,
-    1592,
-    1591,
-    1591,
-    1590,
-    1589,
-    1588,
-    1587,
-    1586,
-    1585,
-    1584,
-    1583,
-    1582,
-    1581,
-    1580,
-    1579,
-    1578,
-    1576,
-    1575,
-    1574,
-    1573,
-    1572,
-    1570,
-    1569,
-    1568,
-    1566,
-    1565,
-    1564,
-    1562,
-    1561,
-    1559,
-    1558,
-    1556,
-    1555,
-    1553,
-    1552,
-    1550,
-    1548,
-    1547,
-    1545,
-    1543,
-    1542,
-    1540,
-    1538,
-    1537,
-    1535,
-    1533,
-    1531,
-    1529,
-    1527,
-    1525,
-    1524,
-    1522,
-    1520,
-    1518,
-    1516,
-    1514,
-    1512,
-    1510,
-    1508,
-    1506,
-    1503,
-    1501,
-    1499,
-    1497,
-    1495,
-    1493,
-    1490,
-    1488,
-    1486,
-    1484,
-    1481,
-    1479,
-    1477,
-    1474,
-    1472,
-    1469,
-    1467,
-    1465,
-    1462,
-    1460,
-    1457,
-    1455,
-    1452,
-    1450,
-    1447,
-    1444,
-    1442,
-    1439,
-    1437,
-    1434,
-    1431,
-    1429,
-    1426,
-    1423,
-    1420,
-    1418,
-    1415,
-    1412,
-    1409,
-    1406,
-    1404,
-    1401,
-    1398,
-    1395,
-    1392,
-    1389,
-    1386,
-    1383,
-    1380,
-    1377,
-    1374,
-    1371,
-    1368,
-    1365,
-    1362,
-    1359,
-    1356,
-    1353,
-    1350,
-    1347,
-    1344,
-    1341,
-    1338,
-    1334,
-    1331,
-    1328,
-    1325,
-    1322,
-    1318,
-    1315,
-    1312,
-    1309,
-    1305,
-    1302,
-    1299,
-    1295,
-    1292,
-    1289,
-    1285,
-    1282,
-    1279,
-    1275,
-    1272,
-    1268,
-    1265,
-    1261,
-    1258,
-    1255,
-    1251,
-    1248,
-    1244,
-    1241,
-    1237,
-    1234,
-    1230,
-    1227,
-    1223,
-    1220,
-    1216,
-    1212,
-    1209,
-    1205,
-    1202,
-    1198,
-    1195,
-    1191,
-    1187,
-    1184,
-    1180,
-    1176,
-    1173,
-    1169,
-    1165,
-    1162,
-    1158,
-    1154,
-    1151,
-    1147,
-    1143,
-    1140,
-    1136,
-    1132,
-    1128,
-    1125,
-    1121,
-    1117,
-    1114,
-    1110,
-    1106,
-    1102,
-    1099,
-    1095,
-    1091,
-    1087,
-    1084,
-    1080,
-    1076,
-    1072,
-    1068,
-    1065,
-    1061,
-    1057,
-    1053,
-    1049,
-    1046,
-    1042,
-    1038,
-    1034,
-    1030,
-    1027,
-    1023,
-    1019,
-    1015,
-    1011,
-    1008,
-    1004,
-    1000,
-    996,
-    992,
-    989,
-    985,
-    981,
-    977,
-    973,
-    970,
-    966,
-    962,
-    958,
-    954,
-    951,
-    947,
-    943,
-    939,
-    935,
-    932,
-    928,
-    924,
-    920,
-    916,
-    913,
-    909,
-    905,
-    901,
-    898,
-    894,
-    890,
-    886,
-    883,
-    879,
-    875,
-    872,
-    868,
-    864,
-    860,
-    857,
-    853,
-    849,
-    846,
-    842,
-    838,
-    835,
-    831,
-    827,
-    824,
-    820,
-    816,
-    813,
-    809,
-    805,
-    802,
-    798,
-    795,
-    791,
-    788,
-    784,
-    780,
-    777,
-    773,
-    770,
-    766,
-    763,
-    759,
-    756,
-    752,
-    749,
-    745,
-    742,
-    739,
-    735,
-    732,
-    728,
-    725,
-    721,
-    718,
-    715,
-    711,
-    708,
-    705,
-    701,
-    698,
-    695,
-    691,
-    688,
-    685,
-    682,
-    678,
-    675,
-    672,
-    669,
-    666,
-    662,
-    659,
-    656,
-    653,
-    650,
-    647,
-    644,
-    641,
-    638,
-    635,
-    632,
-    629,
-    626,
-    623,
-    620,
-    617,
-    614,
-    611,
-    608,
-    605,
-    602,
-    599,
-    596,
-    594,
-    591,
-    588,
-    585,
-    582,
-    580,
-    577,
-    574,
-    571,
-    569,
-    566,
-    563,
-    561,
-    558,
-    556,
-    553,
-    550,
-    548,
-    545,
-    543,
-    540,
-    538,
-    535,
-    533,
-    531,
-    528,
-    526,
-    523,
-    521,
-    519,
-    516,
-    514,
-    512,
-    510,
-    507,
-    505,
-    503,
-    501,
-    499,
-    497,
-    494,
-    492,
-    490,
-    488,
-    486,
-    484,
-    482,
-    480,
-    478,
-    476,
-    475,
-    473,
-    471,
-    469,
-    467,
-    465,
-    463,
-    462,
-    460,
-    458,
-    457,
-    455,
-    453,
-    452,
-    450,
-    448,
-    447,
-    445,
-    444,
-    442,
-    441,
-    439,
-    438,
-    436,
-    435,
-    434,
-    432,
-    431,
-    430,
-    428,
-    427,
-    426,
-    425,
-    424,
-    422,
-    421,
-    420,
-    419,
-    418,
-    417,
-    416,
-    415,
-    414,
-    413,
-    412,
-    411,
-    410,
-    409,
-    409,
-    408,
-    407,
-    406,
-    405,
-    405,
-    404,
-    403,
-    403,
-    402,
-    401,
-    401,
-    400,
-    400,
-    399,
-    399,
-    398,
-    398,
-    397,
-    397,
-    397,
-    396,
-    396,
-    396,
-    395,
-    395,
-    395,
-    395,
-    395,
-    394,
-    394,
-    394,
-    394,
-    394,
-    394,
-    394,
-    394,
-    394,
-    394,
-    394,
-    394,
-    394,
-    395,
-    395,
-    395,
-    395,
-    395,
-    396,
-    396,
-    396,
-    397,
-    397,
-    397,
-    398,
-    398,
-    399,
-    399,
-    400,
-    400,
-    401,
-    401,
-    402,
-    403,
-    403,
-    404,
-    405,
-    405,
-    406,
-    407,
-    408,
-    409,
-    409,
-    410,
-    411,
-    412,
-    413,
-    414,
-    415,
-    416,
-    417,
-    418,
-    419,
-    420,
-    421,
-    422,
-    424,
-    425,
-    426,
-    427,
-    428,
-    430,
-    431,
-    432,
-    434,
-    435,
-    436,
-    438,
-    439,
-    441,
-    442,
-    444,
-    445,
-    447,
-    448,
-    450,
-    452,
-    453,
-    455,
-    457,
-    458,
-    460,
-    462,
-    463,
-    465,
-    467,
-    469,
-    471,
-    473,
-    475,
-    476,
-    478,
-    480,
-    482,
-    484,
-    486,
-    488,
-    490,
-    492,
-    494,
-    497,
-    499,
-    501,
-    503,
-    505,
-    507,
-    510,
-    512,
-    514,
-    516,
-    519,
-    521,
-    523,
-    526,
-    528,
-    531,
-    533,
-    535,
-    538,
-    540,
-    543,
-    545,
-    548,
-    550,
-    553,
-    556,
-    558,
-    561,
-    563,
-    566,
-    569,
-    571,
-    574,
-    577,
-    580,
-    582,
-    585,
-    588,
-    591,
-    594,
-    596,
-    599,
-    602,
-    605,
-    608,
-    611,
-    614,
-    617,
-    620,
-    623,
-    626,
-    629,
-    632,
-    635,
-    638,
-    641,
-    644,
-    647,
-    650,
-    653,
-    656,
-    659,
-    662,
-    666,
-    669,
-    672,
-    675,
-    678,
-    682,
-    685,
-    688,
-    691,
-    695,
-    698,
-    701,
-    705,
-    708,
-    711,
-    715,
-    718,
-    721,
-    725,
-    728,
-    732,
-    735,
-    739,
-    742,
-    745,
-    749,
-    752,
-    756,
-    759,
-    763,
-    766,
-    770,
-    773,
-    777,
-    780,
-    784,
-    788,
-    791,
-    795,
-    798,
-    802,
-    805,
-    809,
-    813,
-    816,
-    820,
-    824,
-    827,
-    831,
-    835,
-    838,
-    842,
-    846,
-    849,
-    853,
-    857,
-    860,
-    864,
-    868,
-    872,
-    875,
-    879,
-    883,
-    886,
-    890,
-    894,
-    898,
-    901,
-    905,
-    909,
-    913,
-    916,
-    920,
-    924,
-    928,
-    932,
-    935,
-    939,
-    943,
-    947,
-    951,
-    954,
-    958,
-    962,
-    966,
-    970,
-    973,
-    977,
-    981,
-    985,
-    989,
-    992,
-    996,
-};
-
